+++ backend/skynet_backend/src/main/java/toctep/skynet/backend/dal/domain/hashtag/IHashtag.java
@@ -2,7 +2,7 @@ package toctep.skynet.backend.dal.domain.hashtag;
 
 import toctep.skynet.backend.dal.domain.IDomain;
 
-public interface IHashtag extends IDomain<Long> {
+public interface IHashtag extends IDomain<Integer> {
 
 	String getText();
